% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babetteinstall.R
\docType{package}
\name{babetteinstall}
\alias{babetteinstall}
\title{\code{babetteinstall}: A package to install BEAST2.}
\description{
\code{babetteinstall} allows to install and uninstall BEAST2,
a popular Bayesian phylogenetics tool, using
an R interface.
}
\details{
\code{babetteinstall} in a non-CRAN extension of the
\code{babette} package. With \code{babette} one can work with BEAST2
using an R interface.
}
